Fix some translation errors
